Make Gergich give better information about yarn.lock differences
Thought it might be helpful for Gergich to provide diffs on the
yarn.lock file so we can have a better idea what we have installed
that doesn't belong in "vanilla" Canvas.

flag=none
test plan
 none, but see the lint-review -2 comment on g/287836
 for an example of what the output looks like

Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/286910
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Jon Scheiding <jon.scheiding@instructure.com>
QA-Review: Gonzalo Penaranda <gonzalo.penaranda@instructure.com>

+++ build/new-jenkins/linters/run-gergich-yarn.sh
@@ -2,15 +2,30 @@
 
 set -ex
 
-read -r -a PLUGINS_LIST_ARR <<< "$PLUGINS_LIST"
-rm -rf $(printf 'gems/plugins/%s ' "${PLUGINS_LIST_ARR[@]}")
+if [ -z "$SKIP_YARN" ]; then
+  read -r -a PLUGINS_LIST_ARR <<< "$PLUGINS_LIST"
+  rm -rf $(printf 'gems/plugins/%s ' "${PLUGINS_LIST_ARR[@]}")
 
-export DISABLE_POSTINSTALL=1
-yarn install --ignore-optional || yarn install --ignore-optional --network-concurrency 1
+  export DISABLE_POSTINSTALL=1
+  yarn install --ignore-optional || yarn install --ignore-optional --network-concurrency 1
+fi
 
 if ! git diff --exit-code yarn.lock; then
+  #
+  # Put yarn.lock diff into the build output
+  #
+  git --no-pager diff yarn.lock
+
+  #
+  # Put yarn.lock diff into the Gergich message
+  #
+  diff="$(git diff yarn.lock)"
+  diff="\n\n\`\`\`\n$diff\n\`\`\`"
+  diff=${diff//$'\n'/'\n'}
+  diff=${diff//$'"'/'\"'}
+
   message="yarn.lock changes need to be checked in. Make sure you run 'yarn install' without private canvas-lms plugins installed."
-  gergich comment "{\"path\":\"yarn.lock\",\"position\":1,\"severity\":\"error\",\"message\":\"$message\"}"
+  gergich comment "{\"path\":\"yarn.lock\",\"position\":1,\"severity\":\"error\",\"message\":\"$message$diff\"}"
 else
   yarn dedupe-yarn
